# api-cms/docs/swagger.yaml
...
info:
  contact: {}
paths:
...
  /api/v1/common/mentions/{mention_id}:
    delete:
      consumes:
      - application/json
      description: Xóa Mention
      parameters:
      - description: mention_id
        in: path
        name: mention_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xóa Mention
      tags:
      - Mention
    get:
      consumes:
      - application/json
      description: Xem chi tiết Mention
      parameters:
      - description: mention_id
        in: path
        name: mention_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem chi tiết Mention
      tags:
      - Mention
  /api/v1/common/mentions/{mention_id}/assign-label:
    put:
      consumes:
      - application/json
      description: Gán nhãn cho Mention
      parameters:
      - description: ID of the Mention
        in: path
        name: mention_id
        required: true
        type: string
      - description: Name of the Label to assign
        in: query
        name: label_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Gán nhãn cho Mention
      tags:
      - Mention
  /api/v1/common/mentions/{mention_id}/assign-label-sentences:
    put:
      consumes:
      - application/json
      description: Gán nhãn cho Sentences
      parameters:
      - description: ID of the Mention
        in: path
        name: mention_id
        required: true
        type: string
      - description: Index of the sentence in the Sentences array
        in: query
        name: sentence_index
        required: true
        type: integer
      - description: Name of the Label to assign
        in: query
        name: label_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Gán nhãn cho Sentences
      tags:
      - Mention
...
  /api/v1/common/mentions/addLabelFace:
    get:
      consumes:
      - application/json
      description: Xem danh sách Mention chưa gán nhãn
      parameters:
      - description: Account ID
        in: query
        name: account_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Xem danh sách Mention Face chưa gán nhãn
      tags:
      - Mention
...
swagger: "2.0"